# Build only test projects and split out building TFMs into separate queues