--- a/src/Test/TestDependencies/TestDependencies-net6.csproj
+++ b/src/Test/TestDependencies/TestDependencies-net6.csproj
@@ -33,6 +33,12 @@
     <ProjectReference Include="..\..\Workspaces\CSharpTest\Microsoft.CodeAnalysis.CSharp.Workspaces.UnitTests.csproj" ReferenceOutputAssembly="false" />
     <ProjectReference Include="..\..\Workspaces\MSBuildTest\Microsoft.CodeAnalysis.Workspaces.MSBuild.UnitTests.csproj" ReferenceOutputAssembly="false" />
     <ProjectReference Include="..\..\Workspaces\Remote\ServiceHubTest\Microsoft.CodeAnalysis.Remote.ServiceHub.UnitTests.csproj" ReferenceOutputAssembly="false" />
+    <ProjectReference Include="..\..\Compilers\VisualBasic\Test\Emit\Microsoft.CodeAnalysis.VisualBasic.Emit.UnitTests.vbproj" ReferenceOutputAssembly="false" />
+    <ProjectReference Include="..\..\Compilers\VisualBasic\Test\IOperation\Roslyn.Compilers.VisualBasic.IOperation.UnitTests.vbproj" ReferenceOutputAssembly="false" />
+    <ProjectReference Include="..\..\Compilers\VisualBasic\Test\Semantic\Microsoft.CodeAnalysis.VisualBasic.Semantic.UnitTests.vbproj" ReferenceOutputAssembly="false" />
+    <ProjectReference Include="..\..\Compilers\VisualBasic\Test\Symbol\Microsoft.CodeAnalysis.VisualBasic.Symbol.UnitTests.vbproj" ReferenceOutputAssembly="false" />
+    <ProjectReference Include="..\..\Compilers\VisualBasic\Test\Syntax\Microsoft.CodeAnalysis.VisualBasic.Syntax.UnitTests.vbproj" ReferenceOutputAssembly="false" />
+    <ProjectReference Include="..\..\Workspaces\VisualBasicTest\Microsoft.CodeAnalysis.VisualBasic.Workspaces.UnitTests.vbproj" ReferenceOutputAssembly="false" />
   </ItemGroup>
   
   <ItemGroup Condition="'$(TargetFramework)' == 'net472'">