As part of @dibarbet's work on running integration tests on helix he also changed things so we only build the root projects that were actually deployed: https://github.com/dotnet/roslyn/pull/63067

This resulted in a significant speedup that I wanted to investigate. After looking at a lot of binlogs most of the speedup is due to the happy accident of us only building the net472 tfm. After experimenting with different approaches, building a TFM per-machine cuts the build time by 10 minutes for integration tests, 5 minutes for windows builds, and halves it in unix builds.

| Definition | Current Time | Time With TFM Split |
|------------|--------------|:---------------------:|
| | <img width="241" alt="image" src="https://user-images.githubusercontent.com/9797472/187611447-666ba62a-6ce6-4388-878e-d355f6e1bdec.png"> | <img width="236" alt="image" src="https://user-images.githubusercontent.com/9797472/187612522-460ee8db-bb8f-4fdc-a543-bbf39a1250fd.png"> <img width="237" alt="image" src="https://user-images.githubusercontent.com/9797472/187613873-5f5d66e3-6890-4031-a601-f0c8e63a858f.png"> |
| | <img width="236" alt="image" src="https://user-images.githubusercontent.com/9797472/187611582-1ee5bed3-a3c8-446e-a446-8ef1b11dd8fb.png"> | <img width="236" alt="image" src="https://user-images.githubusercontent.com/9797472/187612648-c9a446bf-bab2-4ee6-ae2b-fedd41ce626f.png"> <img width="232" alt="image" src="https://user-images.githubusercontent.com/9797472/187614047-6217e3d8-3e1c-45d0-87e3-47eb1cadb777.png"> | 
| | <img width="236" alt="image" src="https://user-images.githubusercontent.com/9797472/187611645-baf3d9ce-5b92-44bc-9035-0b1f1ef891f3.png"> | <img width="244" alt="image" src="https://user-images.githubusercontent.com/9797472/187612796-7e0abbdd-f956-46fd-b636-8a44139fd675.png"> |
| | <img width="248" alt="image" src="https://user-images.githubusercontent.com/9797472/187612333-2c793621-f7a2-4362-a3c2-9dc4e5f6afe1.png"> | ![image](https://user-images.githubusercontent.com/9797472/187741616-5daa88c4-8952-4a8e-bd0a-ab85a5da87bf.png) |
| | <img width="253" alt="image" src="https://user-images.githubusercontent.com/9797472/187612394-15e4f16a-8cca-41cb-a7f6-57758ec5816e.png"> | <!-- in progress --> |
| | <img width="251" alt="image" src="https://user-images.githubusercontent.com/9797472/187612193-e385afd9-b052-4b16-971c-301fe09492e7.png"> | <img width="208" alt="image" src="https://user-images.githubusercontent.com/9797472/187741739-b1cadd4c-e471-4ae6-b896-27c1e8980d11.png">|